Make the SPLL ampl_loop work!! (carefully cosim tested) 1) FA filter is cleaned on reset 2) Reviewed all pipeline formats
We had a very IMPORTANT PROBLEM: the FA had to be correctly reset (including internal FIFO).
Otherwise, RESULTS WERE SHIT.

If ModelSim vsim was restarted, there was no problem and entity seemed to work, as code take care or reinitializing the FIFO memory to 0.
But just using the ampl_loop rst port, things were not working as some state was left inside the MVA.

ampl_loop has an additional state for cleaning the MVA filter!

ALSO, we have to be VERY VERY CAREFUL with pipeline formats. Added documentation and
corrected all the formast. SQRT new prec parametrization functionaly was needed.