Configuration: avoid warnings from preprocessor
On recent `cpp`, When the c preprocessor finds a "'" character in a pure
C or CPP source, it checks if it has been closed. The VHDL source
files' comments are NOT recognized as such by cpp, and therefore
complains that the normal apostrophes in the comments are unclosed
character literals.

Avoid it by changing the "language" to something that makes it not complain